+++ packages/paste-website/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.13.17](https://github.com/twilio-labs/paste/compare/@twilio-paste/website@2.13.16...@twilio-paste/website@2.13.17) (2020-10-29)
+
+**Note:** Version bump only for package @twilio-paste/website
+
+
+
+
+
+## [2.13.16](https://github.com/twilio-labs/paste/compare/@twilio-paste/website@2.13.15...@twilio-paste/website@2.13.16) (2020-10-28)
+
+**Note:** Version bump only for package @twilio-paste/website
+
+
+
+
+
+## [2.13.15](https://github.com/twilio-labs/paste/compare/@twilio-paste/website@2.13.14...@twilio-paste/website@2.13.15) (2020-10-28)
+
+**Note:** Version bump only for package @twilio-paste/website
+
+
+
+
+
+## [2.13.14](https://github.com/twilio-labs/paste/compare/@twilio-paste/website@2.13.13...@twilio-paste/website@2.13.14) (2020-10-27)
+
+**Note:** Version bump only for package @twilio-paste/website
+
+
+
+
+
 ## [2.13.13](https://github.com/twilio-labs/paste/compare/@twilio-paste/website@2.13.12...@twilio-paste/website@2.13.13) (2020-10-27)
 
 **Note:** Version bump only for package @twilio-paste/website